+++ src/app/common/components/dataset-list-component/dataset-list-item/dataset-list-item.component.ts
@@ -18,8 +18,9 @@ import { FeatureFlagDirective } from "../../../directives/feature-flag.directive
 import { DatasetVisibilityComponent } from "../../dataset-visibility/dataset-visibility.component";
 import { RouterLink } from "@angular/router";
 import { MatIconModule } from "@angular/material/icon";
-import { NgIf, NgFor, NgClass } from "@angular/common";
+import { NgIf, NgFor } from "@angular/common";
 import AppValues from "src/app/common/values/app.values";
+import { DatasetKindComponent } from "../../dataset-kind/dataset-kind.component";
 
 @Component({
     selector: "app-dataset-list-item",
@@ -28,7 +29,6 @@ import AppValues from "src/app/common/values/app.values";
     standalone: true,
     imports: [
         //-----//
-        NgClass,
         NgFor,
         NgIf,
         RouterLink,
@@ -42,6 +42,7 @@ import AppValues from "src/app/common/values/app.values";
 
         //-----//
         DatasetVisibilityComponent,
+        DatasetKindComponent,
         DisplayTimeComponent,
         FeatureFlagDirective,
     ],