+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] 
+### Changed
+- Metadata tab: combined "Dataset Type", "Name", and "Owner" into one property "Dataset"
+
+
 ## [0.56.0] - 2025-08-28
 ### Added
 - Visualizing webhook task failures in flow table and history, differentiating between: